+++ packages/browser/src/login/oidc/OidcLoginHandler.ts
@@ -37,11 +37,12 @@ import {
   ILoginHandler,
   IOidcHandler,
   IOidcOptions,
+  IStorageUtility,
 } from "@inrupt/solid-client-authn-core";
 
 import ConfigurationError from "../../errors/ConfigurationError";
-import { IDpopClientKeyManager } from "../../dpop/DpopClientKeyManager";
 import URL from "url-parse";
+import { IClient } from "@inrupt/oidc-dpop-client-browser";
 
 /**
  * @hidden
@@ -49,11 +50,10 @@ import URL from "url-parse";
 @injectable()
 export default class OidcLoginHandler implements ILoginHandler {
   constructor(
+    @inject("storageUtility") private storageUtility: IStorageUtility,
     @inject("oidcHandler") private oidcHandler: IOidcHandler,
     @inject("issuerConfigFetcher")
     private issuerConfigFetcher: IIssuerConfigFetcher,
-    @inject("dpopClientKeyManager")
-    private dpopClientKeyManager: IDpopClientKeyManager,
     @inject("clientRegistrar") private clientRegistrar: IClientRegistrar
   ) {}
 
@@ -81,23 +81,56 @@ export default class OidcLoginHandler implements ILoginHandler {
       options.oidcIssuer as URL
     );
 
+    let dynamicClientRegistration: IClient;
+    if (options.clientId) {
+      dynamicClientRegistration = {
+        clientId: options.clientId,
+        clientSecret: options.clientSecret,
+        clientName: options.clientName,
+      };
+    } else {
+      const clientId = await this.storageUtility.getForUser(
+        "clientApplicationRegistrationInfo", //options.sessionId,
+        "clientId"
+      );
+
+      if (clientId) {
+        dynamicClientRegistration = {
+          clientId,
+          clientSecret: await this.storageUtility.getForUser(
+            "clientApplicationRegistrationInfo",
+            "clientSecret"
+          ),
+          clientName: options.clientName,
+        };
+      } else {
+        dynamicClientRegistration = await this.clientRegistrar.getClient(
+          {
+            sessionId: options.sessionId,
+            clientName: options.clientName,
+            redirectUrl: options.redirectUrl,
+          },
+          issuerConfig
+        );
+
+        await this.storageUtility.setForUser(
+          "clientApplicationRegistrationInfo",
+          {
+            clientId: dynamicClientRegistration.clientId,
+            clientSecret: dynamicClientRegistration.clientSecret as string,
+          }
+        );
+      }
+    }
+
     // Construct OIDC Options
     const OidcOptions: IOidcOptions = {
       issuer: options.oidcIssuer as URL,
       // TODO: differentiate if DPoP should be true
-      dpop: true,
+      dpop: options.tokenType.toLowerCase() === "dpop",
       redirectUrl: options.redirectUrl as URL,
       issuerConfiguration: issuerConfig,
-      client: await this.clientRegistrar.getClient(
-        {
-          sessionId: options.sessionId,
-          clientId: options.clientId,
-          clientSecret: options.clientSecret,
-          clientName: options.clientName,
-          redirectUrl: options.redirectUrl,
-        },
-        issuerConfig
-      ),
+      client: dynamicClientRegistration,
       sessionId: options.sessionId,
       handleRedirect: options.handleRedirect,
     };