+++ packages/browser/src/login/oidc/redirectHandler/AggregateRedirectHandler.ts
@@ -39,7 +39,7 @@ import {
  */
 @injectable()
 export default class AggregateRedirectHandler
-  extends AggregateHandler<[string], ISessionInfo | undefined>
+  extends AggregateHandler<[string], ISessionInfo & { fetch: typeof fetch }>
   implements IRedirectHandler {
   constructor(
     @injectAll("redirectHandlers") redirectHandlers: IRedirectHandler[]